Avoid false positive warning for generic typed method receiver
Because of strange reasons, declared generic instance method receiver is
not the subtype of the class bound. However, our initial goal is just to
ensure main modifier of declared method receiver to be subtype of main
modifier of class bound. Therefore, we minize the scope of the subtyping
check, and leave the problem to CF(actually it isn't a problem on CF
side at all)